Allow optional weights input in kruskal_to_tensor
In some applications it is useful to scale the factor matrices using a vector of
weights. An optional argument `weights` is added to `kruskal_to_tensor`.

+++ tensorly/backend/mxnet_backend.py
@@ -26,7 +26,7 @@ def context(tensor):
     """
     return {'ctx':tensor.context, 'dtype':tensor.dtype}
 
-def tensor(data, ctx=mx.cpu(), dtype="float64"):
+def tensor(data, ctx=mx.cpu(), dtype='float32'):
     """Tensor class
     """
     if dtype is None and isinstance(data, numpy.ndarray):
